+++ frontend/src/components/CardUI.js
@@ -5,18 +5,7 @@ function CardUI()
     var card = '';
     var search = '';
 
-    const app_name = 'facebetter'
-    function buildPath(route)
-    {
-        if (process.env.NODE_ENV === 'production') 
-        {
-            return 'https://' + app_name +  '.herokuapp.com/' + route;
-        }
-        else
-        {        
-            return 'http://localhost:5000/' + route;
-        }
-    }
+    var bp = require('./Path.js');
 
     const [message,setMessage] = useState('');
     const [searchResults,setResults] = useState('');
@@ -37,7 +26,7 @@ function CardUI()
 
         try
         {
-            const response = await fetch(buildPath('api/addcard'),
+            const response = await fetch(bp.buildPath('api/addcard'),
             {method:'POST',body:js,headers:{'Content-Type': 'application/json'}});
 
             var txt = await response.text();
@@ -68,7 +57,7 @@ function CardUI()
 
         try
         {
-            const response = await fetch(buildPath('api/searchcards'),
+            const response = await fetch(bp.buildPath('api/searchcards'),
             {method:'POST',body:js,headers:{'Content-Type': 'application/json'}});
 
             var txt = await response.text();
