+++ frontend/src/components/Path.js
@@ -1,0 +1,13 @@
+const app_name = 'facebetter'
+exports.buildPath = 
+function buildPath(route)
+{
+    if (process.env.NODE_ENV === 'production') 
+    {
+        return 'https://' + app_name +  '.herokuapp.com/' + route;
+    }
+    else
+    {        
+        return 'http://localhost:5000/' + route;
+    }
+}
